requirements.txt
apiclient==1.0.4
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configparser==3.7.4
Django==2.2.3
dropbox==9.4.0
Flask==1.1.1
google-auth==1.6.3
greenlet==0.4.15
httplib2==0.13.0
idna==2.8
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.0.2
opencv-python==4.1.0.25
Pillow==6.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycups==1.9.74
PyDrive==1.3.1
pygame==1.9.6
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
runtime==0.1.4
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
